pulls dual list box into its own component
